shrink scope to not account for mobile, cleaned up logic around authentication
