--- frontend/src/components/EnvironmentSmall.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-// for small screens
-const EnvironmentSmall = props => {
-  return (
-    <div>
-      <p> this is for mobile development, should do it eventually mobile is very popular nowadays</p>
-    </div>
-  )
-}
-
-export default EnvironmentSmall
