+++ frontend/src/components/Header.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import logo from '../logo.svg';
+// import logo from '../logo.svg';
 import '../style/hamburger.css';
 
 const Header = props => {
