+++ frontend/src/components/Home.jsx
@@ -1,9 +1,11 @@
 import React from 'react'
 
-const Home = () => {
+const Home = props => {
+  const {setAuthStatus} = props;
   return (
     <div>
       <h1>this is home page</h1>
+      <button onClick = {()=>{setAuthStatus(false)}}>logout</button>
     </div>
   )
 }
