+++ frontend/src/components/Protected.jsx
@@ -1,0 +1,69 @@
+import React, { useState, useEffect } from 'react';
+import Header from './Header';
+import { Switch, Route } from 'react-router-dom';
+import Home from './Home';
+import NotFound from './NotFound';
+import Menu from './Menu';
+
+const Environment = props => {
+  const [userInfo, setUserInfo] = useState({});
+  const [isOpen, setIsOpen] = useState(false);
+  const { authStatus, setAuthStatus, history } = props;
+
+  // check authentication when authstatus is updated
+  useEffect(() => {
+    if (!authStatus) history.push('/');
+  }, [authStatus]);
+
+  useEffect(() => {
+    // api call for user data
+    //* temporary mock data
+    setUserInfo({
+      userID: 0,
+      name: 'PersonMcPersonFace',
+      classes: {
+        0: {
+          schoolID: 0,
+          classID: 0,
+          sectionID: 0,
+          className: 'classyMcClassface',
+          time: new Date(),
+          professor: 'TeacheryMcTeacherFace',
+        },
+      },
+      picture: '',
+    });
+    console.log('mount');
+    return () => {
+      console.log('cleanup');
+    };
+  }, []);
+
+  return (
+    <>
+      <Header isOpen={isOpen} setIsOpen={setIsOpen} />
+
+      <div className="App__wrapper">
+        <Menu isOpen={isOpen} name={userInfo.name} picture={userInfo.picture} />
+        <div
+          className={`App__container ${
+            isOpen ? 'App__container--open' : 'App__container--closed'
+          }`}>
+          <Switch>
+            <Route
+              path={`/home`}
+              render={props => (
+                <Home {...props} setAuthStatus={setAuthStatus} />
+              )}
+            />
+            <Route path={`/calendar`} component={NotFound} />
+            <Route path={`/thing`} render={props => <NotFound {...props} />} />
+            ))}
+          </Switch>
+        </div>
+      </div>
+    </>
+  );
+};
+
+export default Environment;
